pyproject.toml
[tool.poetry]
name = "ty-fastapi-jwt-auth-server"
version = "0.1.0"
description = ""
authors = ["TY <tyyyy66@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
python-jose = "^3.4.0"
python-multipart = "^0.0.20"
pydantic-settings = "^2.8.1"
fastapi-cli = "^0.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
